Check the correct integer type for tastudio.setplayback
New lua actually has integers so this is a long instead of a double

+++ src/BizHawk.Client.EmuHawk/tools/Lua/Libraries/TAStudioLuaLibrary.cs
@@ -139,7 +139,7 @@ namespace BizHawk.Client.EmuHawk
 				_luaLibsImpl.IsUpdateSupressed = true;
 
 				int f;
-				if (frame is double frameNumber)
+				if (frame is long frameNumber)
 				{
 					f = (int)frameNumber;
 				}
